Add frontend ECS and simplify infrastructure
- Add frontend ECS service (3GB, Next.js) sharing VPC with backend
- Rename ecs.ts → backend-ecs.ts for clarity
- Align DynamoDB: reduce 43 tables to 6 core tables matching dynamo_schemas.py
- Comment out unused constructs (Lambda, API Gateway, Cognito) for MVP
- Add dynamodb-admin GUI to local deployment on localhost:8001
